feat: Migrate to gemmi backend as default with legacy archive for compatibility
- Migrated from pure Python to gemmi C++ backend as default for 10x+ performance improvement
- Archived original implementation to sloth.legacy for full backward compatibility
- Removed use_gemmi parameter - gemmi is now the default backend
- Updated all tests to match gemmi output format and validate both implementations
- Made gemmi a required dependency (was optional) in pyproject.toml
- Enhanced version management with dynamic versioning and migration info
- Updated README with comprehensive migration guide and version compatibility checks
- Verified 83/83 tests passing with full feature parity across backends

BREAKING CHANGE: use_gemmi parameter removed (gemmi now default)
Migration: Use sloth.legacy.LegacyMMCIFParser for original implementation
