+++ pyproject.toml
@@ -31,6 +31,7 @@ classifiers = [
 keywords = ["mmcif", "cif", "crystallography", "structural-biology", "parser", "pdb"]
 dependencies = [
     "msgpack",
+    "gemmi>=0.6.0",
 ]
 requires-python = ">=3.8"
